Tareas/Tarea 5/Gestion Eventos(Python)/requirements.txt
Django==5.2.7
djangorestframework==3.15.1
PyJWT==2.8.0
cryptography==42.0.5
